+++ Sources/SwifterSwift/UIKit/UIImageExtensions.swift
@@ -309,6 +309,15 @@ public extension UIImage {
         self.init(cgImage: aCgImage)
     }
 
+    /// SwifterSwift: Create a new image from a base 64 string.
+    ///
+    /// - Parameters:
+    ///   - base64String: a base-64 `String`, representing the image
+    convenience init?(base64String: String) {
+        guard let data = Data(base64Encoded: base64String) else { return nil }
+        self.init(data: data)
+    }
+
 }
 
 #endif